Revert "Revert "[KBV-399] Refactor to remove address specific functionality""
