+++ accesstoken/src/main/java/uk/gov/di/ipv/cri/common/api/handler/exceptions/ClientAuthenticationException.java
@@ -1,0 +1,11 @@
+package uk.gov.di.ipv.cri.common.api.handler.exceptions;
+
+public class ClientAuthenticationException extends Exception {
+    public ClientAuthenticationException(String message) {
+        super(message);
+    }
+
+    public ClientAuthenticationException(Throwable e) {
+        super(e);
+    }
+}
