+++ build.gradle
@@ -1,0 +1,25 @@
+plugins {
+	id "java"
+	id "com.diffplug.spotless" version "6.1.0"
+}
+
+repositories {
+	maven {
+		url 'https://gds.jfrog.io/artifactory/di-allowed-repos'
+	}
+}
+
+spotless {
+	java {
+		target "**/src/**/*.java"
+		googleJavaFormat("1.13.0").aosp()
+		importOrder "", "javax", "java", "\\#"
+		endWithNewline()
+	}
+	groovyGradle {
+		target '**/*.gradle'
+		greclipse()
+		trimTrailingWhitespace()
+		endWithNewline()
+	}
+}
