+++ jwkset/src/main/java/uk/gov/di/ipv/cri/common/api/handler/EnvironmentService.java
@@ -1,0 +1,9 @@
+package uk.gov.di.ipv.cri.common.api.handler;
+
+import java.util.Optional;
+
+public class EnvironmentService {
+    public String getEnvironmentVariableOrThrow(String envVar) {
+        return Optional.ofNullable(System.getenv(envVar)).orElseThrow();
+    }
+}
