+++ session/build.gradle
@@ -5,7 +5,8 @@ plugins {
 }
 
 dependencies {
-	implementation project(":lib"),
+	implementation project(":common-lib"),
+			configurations.aws,
 			configurations.lambda,
 			configurations.nimbus,
 			configurations.kms,
