--- session/src/main/java/uk/gov/di/ipv/cri/address/api/service/KMSRSADecrypter.java
@@ -1,98 +1,0 @@
-package uk.gov.di.ipv.cri.address.api.service;
-
-import com.nimbusds.jose.EncryptionMethod;
-import com.nimbusds.jose.JOSEException;
-import com.nimbusds.jose.JWEAlgorithm;
-import com.nimbusds.jose.JWEDecrypter;
-import com.nimbusds.jose.JWEHeader;
-import com.nimbusds.jose.crypto.impl.AlgorithmSupportMessage;
-import com.nimbusds.jose.crypto.impl.ContentCryptoProvider;
-import com.nimbusds.jose.jca.JWEJCAContext;
-import com.nimbusds.jose.util.Base64URL;
-import software.amazon.awssdk.core.SdkBytes;
-import software.amazon.awssdk.services.kms.KmsClient;
-import software.amazon.awssdk.services.kms.model.DecryptRequest;
-import software.amazon.awssdk.services.kms.model.DecryptResponse;
-import software.amazon.awssdk.services.kms.model.EncryptionAlgorithmSpec;
-
-import javax.crypto.SecretKey;
-import javax.crypto.spec.SecretKeySpec;
-
-import java.util.Objects;
-import java.util.Set;
-
-class KMSRSADecrypter implements JWEDecrypter {
-    private static final Set<JWEAlgorithm> SUPPORTED_ALGORITHMS = Set.of(JWEAlgorithm.RSA_OAEP_256);
-    private static final Set<EncryptionMethod> SUPPORTED_ENCRYPTION_METHODS =
-            Set.of(EncryptionMethod.A256GCM);
-
-    private final JWEJCAContext jcaContext;
-    private final KmsClient kmsClient;
-    private final String keyId;
-
-    KMSRSADecrypter(String keyId) {
-        this(keyId, KmsClient.builder().build());
-    }
-
-    KMSRSADecrypter(String keyId, KmsClient kmsClient) {
-        this.keyId = keyId;
-        this.kmsClient = kmsClient;
-        this.jcaContext = new JWEJCAContext();
-    }
-
-    @Override
-    public byte[] decrypt(
-            JWEHeader header,
-            Base64URL encryptedKey,
-            Base64URL iv,
-            Base64URL cipherText,
-            Base64URL authTag)
-            throws JOSEException {
-        // Validate required JWE parts
-        if (Objects.isNull(encryptedKey)) {
-            throw new JOSEException("Missing JWE encrypted key");
-        }
-
-        if (Objects.isNull(iv)) {
-            throw new JOSEException("Missing JWE initialization vector (IV)");
-        }
-
-        if (Objects.isNull(authTag)) {
-            throw new JOSEException("Missing JWE authentication tag");
-        }
-
-        JWEAlgorithm alg = header.getAlgorithm();
-
-        if (!SUPPORTED_ALGORITHMS.contains(alg)) {
-            throw new JOSEException(
-                    AlgorithmSupportMessage.unsupportedJWEAlgorithm(alg, supportedJWEAlgorithms()));
-        }
-
-        DecryptRequest decryptRequest =
-                DecryptRequest.builder()
-                        .encryptionAlgorithm(EncryptionAlgorithmSpec.RSAES_OAEP_SHA_256)
-                        .ciphertextBlob(SdkBytes.fromByteArray(encryptedKey.decode()))
-                        .keyId(this.keyId)
-                        .build();
-        DecryptResponse decryptResponse = this.kmsClient.decrypt(decryptRequest);
-        SecretKey cek = new SecretKeySpec(decryptResponse.plaintext().asByteArray(), "AES");
-
-        return ContentCryptoProvider.decrypt(
-                header, encryptedKey, iv, cipherText, authTag, cek, getJCAContext());
-    }
-
-    @Override
-    public Set<JWEAlgorithm> supportedJWEAlgorithms() {
-        return SUPPORTED_ALGORITHMS;
-    }
-
-    @Override
-    public Set<EncryptionMethod> supportedEncryptionMethods() {
-        return SUPPORTED_ENCRYPTION_METHODS;
-    }
-
-    @Override
-    public JWEJCAContext getJCAContext() {
-        return jcaContext;
-    }
-}
