+++ settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name = "di-ipv-cri-common-api"
